requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.3.0
cached-property==1.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Django==3.2.23
django-ckeditor==6.3.2
django-colorfield==0.10.1
django-cors-headers==4.1.0
django-crontab==0.7.1
django-dbbackup==4.0.2
django-jalali==6.0.1
django-jalali-date==1.0.2
django-js-asset==2.0.0
django-mptt==0.14.0
django-redis==5.4.0
django-referrer-policy==1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.5
importlib-metadata==6.7.0
isodate==0.6.1
isort==5.11.5
jdatetime==4.1.1
kavenegar==1.1.2
lxml==4.9.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==9.5.0
platformdirs==4.0.0
psycopg==3.1.13
psycopg2-binary==2.9.9
python-dotenv==0.21.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typed-ast==1.5.5
typing_extensions==4.7.1
urllib3==2.0.7
zeep==4.2.1
zipp==3.15.0
